requirements.txt
google-cloud-speech
google-cloud-aiplatform
requests
